requirements.txt
numpy
Pillow
scipy
tk